+++ CoreAnalyticView/src/au/gov/asd/tac/constellation/views/analyticview/analytics/FactAnalyticPlugin.java
@@ -76,7 +76,7 @@ public abstract class FactAnalyticPlugin extends AnalyticPlugin<FactResult> {
     }
 
     protected final StoreGraph getSubgraph(final GraphWriteMethods graph, final SchemaFactory schemaFactory, final Set<SchemaTransactionType> subgraphTransactionTypes) {
-        return SubgraphUtilities.getSubgraph(graph, schemaFactory.createSchema(), subgraphTransactionTypes, false);
+        return SubgraphUtilities.getTransactionTypeSubgraph(graph, schemaFactory.createSchema(), subgraphTransactionTypes, false);
     }
 
     protected final void copySubgraphToGraph(final GraphWriteMethods graph, final StoreGraph subgraph) {